requirements.txt
pytest=7.4.3
selene>=2.0.b4
allure-pytest=2.13.2